samples/computing-providers/requirements.txt
swan-sdk==0.0.7
python-dotenv==1.0.1